requirements.txt
flask>=3.0.0
pandas>=2.2.2
numpy>=1.26.0
scikit-learn>=1.5.0
xgboost>=2.0.0
shap>=0.45.0
joblib>=1.3.0
pyarrow>=15.0.0
pyyaml>=6.0.1
streamlit>=1.36.0
structlog>=24.1.0
python-json-logger>=2.0.7
prometheus-client>=0.20.0
langchain>=0.2.0
openai>=1.30.0
pytest>=8.2.0
pytest-cov>=5.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.32.0
